+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.18.14
+### is.codion.common.core
+- PropertyStore, property access now based on predicates instead of key prefixes.
 
 ## 0.18.13
 ### is.codion